linux-version: Fix sorting of version strings containing -trunk (Closes: #761614)
The regex used to split up versions into components would yield
'-trunk-flavour' as a single component, but we only treated '-rc' and
'-trunk' as special.  This didn't affect '-rc1-flavour' because it
would break at the digit.

Change it to break at hyphens as well as digits.

Add regression tests for this.

svn path=/dists/trunk/linux-base/; revision=21848