Add linux-update-symlinks command for use by package maintainer scripts
This is simpler than the current logic in the linux maintainer
scripts:

- Now that I know that hard-linking and copying to the default locations
  were broken, I only need to worry about making symlinks
- The logic for updating on installation/upgrade and removal is kept
  together rather than divided across two scripts
- It uses existing functions in the DebianLinux module

But it's also smarter: it builds a list of versions and the files that
belong to them, prioritised based on the current type of change, the
existing symlinks and version sorting.  So it can always point the
symlinks somewhere sensible, unless the last image is being removed.