linux-version: Fix comparing of versions provided via stdin
Currently the trailing "\n" is used in comparisons. I found this when
comparing "5.8.0-50-generic" with "5.8.0-50-generic-64k". Since "\n"
sorts before "-", 5.8.0-50-generic was sorting greater. So, one day
short of the 10th anniversary of the last commit to this tool, I present
this fix. (Closes: #987851)

Note: flash-kernel uses stdin linux-version comparisons to determine if
the kernel is new enough to flash, so devices that use it would be impacted.

Signed-off-by: dann frazier <dannf@debian.org>