Remove perf wrappers, as linux-perf now provides an unversioned perf command
Starting with version 5.16.2-1~exp1, linux-perf is a normal package
containing an unversioned perf executable.  It diverts our wrappers
to avoid conflicts.

We want to avoid the situation that linux-base is upgraded, removing
the perf wrappers, but users only have an old version of linux-perf
installed with a version-suffixed perf command.

A versioned Breaks will cause linux-perf to be upgraded first, which
will install the unversioned perf executable, bash completions, and
man page.  When linux-base is upgraded the wrappers will be removed.
At no point will the perf command be missing.

This won't help users that installed e.g. linux-perf-5.10 without the
metapackage, but there's not much we can realistically do about that.

Closes: #983314, #990165